feat(sidebar): add following list and update layout nomenclature
